Handle unauthorized edit attempts
After prohibiting editing user data for users not logged in or logged
into a different account, this commit aims to handle those attempts
nicely.
Such attempts may occur if the user leaves browser tabs open with
editing tools on them and logs out in a different tab.
Clicking the Edit/Save/Cancel buttons now won't simply break the code.
Editor.js gets the authorization via AJAX response and if the operation
is not authorized, it reloads the page with a respective flash message
and without operation buttons.