Correct instrument update with foreign key issue
If only the instrument's creator had repertoire
entry with the specific instrument, updating the
instrument caused an IntegrityError due to the
foreign key constraint between the instruments and
works tables.
In this case now a new instrument will be created,
the works get updated and the old instrument will be
deleted eventually.